@@ -7,17 +7,23 @@ import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { LogInComponent } from './components/registration/log-in/log-in.component';
 import { SignUpComponent } from './components/registration/sign-up/sign-up.component';
-import { HttpClientModule } from '@angular/common/http';
+import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
 import { HomeComponent } from './components/home/home.component';
 import {AccordionModule} from 'primeng/accordion';     //accordion and accordion tab
 import { CarouselModule } from 'primeng/carousel';
 import { SidebarModule } from 'primeng/sidebar';
+import { SliderComponent } from './components/slider/slider.component';
+import { AuthService } from './services/auth.service';
+import { AuthInterceptor } from './interceptors/auth.interceptor';
+import { MoviesSectionComponent } from './components/movies-section/movies-section.component';
 @NgModule({
   declarations: [
     AppComponent,
     LogInComponent,
     SignUpComponent,
-    HomeComponent
+    HomeComponent,
+    SliderComponent,
+    MoviesSectionComponent
   ],
   imports: [
     BrowserModule,
@@ -30,7 +36,12 @@ import { SidebarModule } from 'primeng/sidebar';
     CarouselModule,
     SidebarModule
   ],
-  providers: [],
+  providers: [ AuthService , 
+    {
+     provide: HTTP_INTERCEPTORS,
+     useClass: AuthInterceptor,
+     multi: true
+    }],
   bootstrap: [AppComponent]
 })
 export class AppModule { }
