@@ -3,12 +3,13 @@ import { RouterModule, Routes } from '@angular/router';
 import { HomeComponent } from './components/home/home.component';
 import { LogInComponent } from './components/registration/log-in/log-in.component';
 import { SignUpComponent } from './components/registration/sign-up/sign-up.component';
+import { AuthGuard } from './Guards/auth.guard';
 
 const routes: Routes = [
   { path: '', pathMatch: 'full', redirectTo: 'home' },
   { path: 'login', component: LogInComponent },
   { path: 'sign-up', component: SignUpComponent },
-  { path: 'home', component: HomeComponent }
+  { path: 'home', component: HomeComponent, canActivate: [AuthGuard] },
 
 ];
 
